drivers: usb_dc_rpi_pico: Implemented vbus detection handling
As per USB 2.0 specification 7.1.5.1: The voltage source on the pull-up
resistor must be derived from or controlled by the power supplied on the
USB cable such that when VBUS is removed, the pull-up resistor does not
supply current on the data line to which it is attached.

This is only relevant for self powered devices. When the device is bus
powered VBUS detection can be left unconnected and the pull-up resistor
is set during init. For self powered devices the pull-up resistor
needs to be set when vbus is attached and cleared when vbus is detached.
Otherwise some systems may not configure the usb device.

Signed-off-by: Lukas Gehreke <lk.gehreke@gmail.com>