FIXUP: switch to using async UART interface
The interface is a bit awkward to use but I suppose works. Tempted to
wrap this in RTIO so I don't have to deal with it and instead can use a
asynchronous packet interface instead but... that would still entail
implementing something on the async interface.

The main issue is that the async callbacks need to signal to the caller
whats going on in some manner. For receive the issue will be the
callback being done in the context of an ISR as per the usual with
Zephyr callback interfaces. This inevitably leads to confusion.

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>